Fix MCP edit page lint layout
